requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
dnspython==2.8.0
email-validator==2.3.0
groq==1.1.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-decouple==3.8
python-dotenv==1.2.2
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
dj-database-url==3.1.2
Django==6.0.3
email-validator==2.3.0
groq==1.1.1
gunicorn==25.1.0
httpx==0.28.1
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.2
dj-database-url==3.1.2
Django==6.0.3
email-validator==2.3.0
groq==1.1.1
gunicorn==25.1.0
httpx==0.28.1
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.2
